Refactored docker commands into CakeDockerPlugin
